+++ conf/evolutions/default/1.sql
@@ -59,7 +59,7 @@ create table invoice_part (
   id                            integer auto_increment not null,
   name                          TEXT not null,
   unit                          varchar(255) not null,
-  quantity                      float not null,
+  quantity                      integer not null,
   unit_price                    integer not null,
   total                         integer not null,
   invoice_id                    integer,
